requirements.txt
llama_index
uvicorn
fastapi
python-dotenv